refactor game manager
